requirements.txt
PySide6
pytesseract
pillow
pdf2image 